Examples: Calculate type sizes for subclassing correctly now that `NewTypeQuery` uses `uint32`

+++ examples/myapp-gnome-meson/src/main.go
@@ -28,5 +28,5 @@ func main() {
 		"flags", gio.GApplicationDefaultFlagsValue,
 	)
 
-	os.Exit(app.Run(len(os.Args), os.Args))
+	os.Exit(int(app.Run(int32(len(os.Args)), os.Args)))
 }
